Fix: move the small problem/idea map after geocode
Because the small maps do not react to the large map anymore (see
previous bug), and only the large map was epxlicitly updated after a
geocode, the small map remained in place. The map-move method now tries
to move the small maps first (which will cause the large map to move as
well).